Each lib includes internally its own deps/redis/*.o
