+++ deps/redis/util.h
@@ -2,7 +2,6 @@
 #define LIBRDB_UTIL_H
 
 #include <stdint.h>
-#include "librdb-hidden-api.h"
 
 #ifdef __GNUC__
 #define likely(x)       __builtin_expect(!!(x), 1)
@@ -20,7 +19,7 @@
 #define MAX_D2STRING_CHARS 128
 
 int string2ll(const char *s, size_t slen, long long *value);
-_LIBRDB_HIDDEN_API int ll2string(char *s, size_t len, long long value);
+int ll2string(char *s, size_t len, long long value);
 int ull2string(char *s, size_t len, unsigned long long value);
 int lpStringToInt64(const char *s, unsigned long slen, int64_t *value);
 unsigned int getEnvVar(const char* varName, unsigned int defaultVal);
@@ -30,6 +29,6 @@ unsigned int getEnvVar(const char* varName, unsigned int defaultVal);
  * incompatibilities as you can also convert double to string that results in
  * loss of precision, or it might not represent inf, -inf or nan values
  * similar to this function output. */
-_LIBRDB_HIDDEN_API int d2string(char *buf, size_t len, double value);
+int d2string(char *buf, size_t len, double value);
 
 #endif /*LIBRDB_UTIL_H*/
